auth-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
authlib
python-jose[cryptography]
python-dotenv